challenge/requirements.txt
openai==0.28
requests
beautifulsoup4
python-dotenv